backend/requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.2.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.4.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
zope.interface==7.2
